veriplagio/requirements.txt
Flask
requests
python-dotenv
PyMuPDF
python-docx
werkzeug
